merge old master with latest shit
this is what's actually what, been running on the door all year
i don't know what's going on with this branch. nobody's touched it in ages. were marc's changes irrelevant? did he merge the ones that matter without explicitly saying so?
making this commit so both branches will merge cleanly
so we can move on from this nonsense but still preserve it for posterity